fix(cli): Fix `tauri init`
Switch to tauri lib instead of `require` to import toml
